+++ e2e/tests/completions/test.js
@@ -7,15 +7,20 @@ const { openMockFile, getFirstResponseOfType, getResponsesOfType } = require('..
 
 const mockFileName = 'main.ts';
 
+const createServerWithMockFile = (fileContents) => {
+    const server = createServer(__dirname, 'echo-plugin');
+    openMockFile(server, mockFileName, fileContents);
+    return server;
+}
+
 describe('Completions', () => {
     it('should return completions inside tagged, single line template', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'const q = test`abcdefg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 16, line: 1, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 18, line: 1, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 23, line: 1, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            'const q = test`abcdefg`',
+            { offset: 16, line: 1 },
+            { offset: 18, line: 1 },
+            { offset: 23, line: 1 }
+        ).then(server => {
             const completionsResponses = getResponsesOfType('completions', server);
             assert.strictEqual(completionsResponses.length, 3);
 
@@ -37,14 +42,12 @@ describe('Completions', () => {
     });
 
     it('should not return completions before tagged template', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'const q = test`abcdefg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 1, line: 1, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 15, line: 1, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            'const q = test`abcdefg`',
+            { offset: 1, line: 1 },
+            { offset: 15, line: 1 }
+        ).then(server => {
             const completionsResponses = getResponsesOfType('completions', server);
-
             for (const response of completionsResponses) {
                 assert.isTrue(response.success);
                 assert.isFalse(response.body.some(item => item.kindModifiers === 'echo'));
@@ -53,11 +56,10 @@ describe('Completions', () => {
     });
 
     it('should return completions for template tag ending with tag name', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'const q = this.is().a.test`abcdefg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 30, line: 1, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            'const q = this.is().a.test`abcdefg`',
+            { offset: 30, line: 1 }
+        ).then(server => {
             const response = getFirstResponseOfType('completions', server);
             assert.isTrue(response.success);
             assert.strictEqual(response.body.length, 1);
@@ -67,26 +69,25 @@ describe('Completions', () => {
     });
 
     it('should not return completions for non-tagged template', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'const q = `abcdefg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 12, line: 1, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            'const q = `abcdefg`',
+            { offset: 12, line: 1 }
+        ).then(server => {
             const response = getFirstResponseOfType('completions', server);
-            assert.isFalse(response.success);
+            assert.isTrue(response.success);
+            assert.strictEqual(response.body.length, 0);
         });
     });
 
     it('should return completions for multiline strings', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, [
-            'const q = test`abc',
-            'cdefg`'
-        ].join('\n'));
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 1, line: 2, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 3, line: 2, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            [
+                'const q = test`abc',
+                'cdefg`'
+            ].join('\n'),
+            { offset: 1, line: 2 },
+            { offset: 3, line: 2 },
+        ).then(server => {
             const completionsResponses = getResponsesOfType('completions', server);
 
             assert.strictEqual(completionsResponses.length, 2);
@@ -100,13 +101,12 @@ describe('Completions', () => {
     });
 
     it('should ignore placeholders in string', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'test`abc${123}e${4}${5}fg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 8, line: 1, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 16, line: 1, prefix: '' } });
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 25, line: 1, prefix: '' } });
-
-        return server.close().then(() => {
+        return getCompletionsInMockFile(
+            'test`abc${123}e${4}${5}fg`',
+            { offset: 8, line: 1 },
+            { offset: 16, line: 1 },
+            { offset: 25, line: 1 }
+        ).then(server => {
             const responses = getResponsesOfType('completions', server);
             assert.strictEqual(responses.length, 3);
 
@@ -127,17 +127,63 @@ describe('Completions', () => {
         });
     });
 
-    it('should allow tag to have space after it', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'const q = test            `abcdefg`');
-        server.send({ command: 'completions', arguments: { file: mockFileName, offset: 30, line: 1, prefix: '' } });
+    it('should allow tag to have space after it', async () => {
+        const server = await getCompletionsInMockFile(
+            'const q = test            `abcdefg`',
+            { offset: 30, line: 1 }
+        );
 
-        return server.close().then(() => {
-            const response = getFirstResponseOfType('completions', server);
-            assert.isTrue(response.success);
-            assert.strictEqual(response.body.length, 1);
-            assert.strictEqual(response.body[0].name, 'ab');
-            assert.strictEqual(response.body[0].kindModifiers, 'echo');
-        });
+        const response = getFirstResponseOfType('completions', server);
+        assert.isTrue(response.success);
+        assert.strictEqual(response.body.length, 1);
+        assert.strictEqual(response.body[0].name, 'ab');
+        assert.strictEqual(response.body[0].kindModifiers, 'echo');
+    });
+
+    it('should allow tag to be a function call', async () => {
+        const server = await getCompletionsInMockFile(
+            'const q = test("bla")`abcdefg`',
+            { offset: 25, line: 1 }
+        );
+
+        const response = getFirstResponseOfType('completions', server);
+        assert.isTrue(response.success);
+        assert.strictEqual(response.body.length, 1);
+        assert.strictEqual(response.body[0].name, 'ab');
+        assert.strictEqual(response.body[0].kindModifiers, 'echo');
+    });
+
+    it('should allow tag to have template parameters', async () => {
+        const server = await getCompletionsInMockFile(
+            'const q = test<number>`abcdefg`',
+            { offset: 26, line: 1 }
+        );
+        const response = getFirstResponseOfType('completions', server);
+        assert.isTrue(response.success);
+        assert.strictEqual(response.body.length, 1);
+        assert.strictEqual(response.body[0].name, 'ab');
+        assert.strictEqual(response.body[0].kindModifiers, 'echo');
+    });
+
+    it('should allow call tag to have template parameters', async () => {
+        const server = await getCompletionsInMockFile(
+            "const q = test<number>('')`abcdefg`",
+            { offset: 30, line: 1 }
+        );
+        const response = getFirstResponseOfType('completions', server);
+        assert.isTrue(response.success);
+        assert.strictEqual(response.body.length, 1);
+        assert.strictEqual(response.body[0].name, 'ab');
+        assert.strictEqual(response.body[0].kindModifiers, 'echo');
     });
 });
+
+    function getCompletionsInMockFile(contents, ...locations) {
+        const server = createServerWithMockFile(contents);
+
+        for (const location of locations) {
+            server.send({ command: 'completions', arguments: { file: mockFileName, ...location } });
+        }
+
+        return server.close().then(() => server);
+    }