+++ e2e/tests/errors/test.js
@@ -1,17 +1,20 @@
 //@ts-check
 const assert = require('chai').assert;
-const { assertRange } = require("../../_assert");
+const { assertRange } = require('../../_assert');
 const createServer = require('../../_server');
-const { openMockFile, getFirstResponseOfType, getResponsesOfType } = require('../../_helpers');
+const { openMockFile, getFirstResponseOfType } = require('../../_helpers');
 
 const mockFileName = 'main.ts';
 
-
+const createServerWithMockFile = (fileContents) => {
+    const server = createServer(__dirname, 'eerror-plugin');
+    openMockFile(server, mockFileName, fileContents);
+    return server;
+}
 
 describe('Errors', () => {
     it('should return errors for single line template string ', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'function test(x) { return x; }; const q = test`eXeeXe`');
+        const server = createServerWithMockFile('function etag(x) { return x; }; const q = etag`eXeeXe`');
         server.send({ command: 'semanticDiagnosticsSync', arguments: { file: mockFileName } });
 
         return server.close().then(() => {
@@ -32,8 +35,7 @@ describe('Errors', () => {
     });
 
     it('should not return errors in placeholders', () => {
-        const server = createServer(__dirname);
-        openMockFile(server, mockFileName, 'function test(x, y) { return x; }; const q = test`e${1}e`');
+        const server = createServerWithMockFile('function etag(x, y) { return x; }; const q = etag`e${1}e`');
         server.send({ command: 'semanticDiagnosticsSync', arguments: { file: mockFileName } });
 
         return server.close().then(() => {