+++ src/standard-template-source-helper.ts
@@ -7,6 +7,70 @@ import ScriptSourceHelper from './script-source-helper';
 import { isTaggedLiteral, isTagged, relative } from './nodes';
 import TemplateContext from './template-context';
 import TemplateSettings from './template-settings';
+import Logger from './logger';
+import { memoize } from './util/memoize';
+
+class PlaceholderSubstituter {
+    public static replacePlaceholders(
+        typescript: typeof ts,
+        settings: TemplateSettings,
+        node: ts.TemplateExpression | ts.NoSubstitutionTemplateLiteral
+    ): string {
+        const literalContents = node.getText().slice(1, -1);
+        if (node.kind === typescript.SyntaxKind.NoSubstitutionTemplateLiteral) {
+            return literalContents;
+        }
+
+        return PlaceholderSubstituter.getSubstitutions(
+            settings,
+            literalContents,
+            PlaceholderSubstituter.getPlaceholderSpans(node));
+    }
+
+    private static getPlaceholderSpans(node: ts.TemplateExpression) {
+        const spans: Array<{ start: number, end: number }> = [];
+        const stringStart = node.getStart() + 1;
+
+        let nodeStart = node.head.end - stringStart - 2;
+        for (const child of node.templateSpans.map(x => x.literal)) {
+            const start = child.getStart() - stringStart + 1;
+            spans.push({ start: nodeStart, end: start });
+            nodeStart = child.getEnd() - stringStart - 2;
+        }
+        return spans;
+    }
+
+    private static getSubstitutions(
+        settings: TemplateSettings,
+        contents: string,
+        locations: ReadonlyArray<{ start: number, end: number }>
+    ): string {
+        if (settings.getSubstitutions) {
+            return settings.getSubstitutions(contents, locations);
+        }
+
+        const parts: string[] = [];
+        let lastIndex = 0;
+        for (const span of locations) {
+            parts.push(contents.slice(lastIndex, span.start));
+            parts.push(this.getSubstitution(settings, contents, span.start, span.end));
+            lastIndex = span.end;
+        }
+        parts.push(contents.slice(lastIndex));
+        return parts.join('');
+    }
+
+    private static getSubstitution(
+        settings: TemplateSettings,
+        templateString: string,
+        start: number,
+        end: number
+    ): string {
+        return settings.getSubstitution
+            ? settings.getSubstitution(templateString, start, end)
+            : 'x'.repeat(end - start);
+    }
+}
 
 class StandardTemplateContext implements TemplateContext {
     constructor(
@@ -14,7 +78,7 @@ class StandardTemplateContext implements TemplateContext {
         public readonly fileName: string,
         public readonly node: ts.TemplateLiteral,
         private readonly helper: ScriptSourceHelper,
-        private readonly templateStringSettings: TemplateSettings
+        private readonly templateSettings: TemplateSettings
     ) { }
 
     public toOffset(position: ts.LineAndCharacter): number {
@@ -29,39 +93,27 @@ class StandardTemplateContext implements TemplateContext {
         return relative(this.stringBodyPosition, docPosition);
     }
 
+    @memoize
     private get stringBodyOffset(): number {
         return this.node.getStart() + 1;
     }
 
+    @memoize
     private get stringBodyPosition(): ts.LineAndCharacter {
         return this.helper.getLineAndChar(this.fileName, this.stringBodyOffset);
     }
 
+    @memoize
     public get text(): string {
-        const literalContents = this.node.getText().slice(1, -1);
-        if (this.node.kind === this.typescript.SyntaxKind.NoSubstitutionTemplateLiteral) {
-            return literalContents;
-        }
-
-        const stringStart = this.node.getStart() + 1;
-        let contents = literalContents;
-        let nodeStart = this.node.head.end - stringStart - 2;
-        for (const child of this.node.templateSpans.map(x => x.literal)) {
-            const start = child.getStart() - stringStart + 1;
-            contents = contents.substr(0, nodeStart) + this.getSubstitution(literalContents, nodeStart, start) + contents.substr(start);
-            nodeStart = child.getEnd() - stringStart - 2;
-        }
-        return contents;
+        return PlaceholderSubstituter.replacePlaceholders(
+            this.typescript,
+            this.templateSettings,
+            this.node);
     }
 
-    private getSubstitution(
-        templateString: string,
-        start: number,
-        end: number
-    ): string {
-        return this.templateStringSettings.getSubstitution
-            ? this.templateStringSettings.getSubstitution(templateString, start, end)
-            : 'x'.repeat(end - start);
+    @memoize
+    public get rawText() {
+        return this.node.getText().slice(1, -1);
     }
 }
 
@@ -69,7 +121,8 @@ export default class StandardTemplateSourceHelper implements TemplateSourceHelpe
     constructor(
         private readonly typescript: typeof ts,
         private readonly templateStringSettings: TemplateSettings,
-        private readonly helper: ScriptSourceHelper
+        private readonly helper: ScriptSourceHelper,
+        private readonly logger: Logger // tslint:disable-line
     ) { }
 
     public getTemplate(