+++ src/template-language-service.ts
@@ -44,12 +44,21 @@ export default interface TemplateLanguageService {
         context: TemplateContext,
         start: number,
         end: number,
-        errorCodes: number[],
+        errorCodes: ReadonlyArray<number>,
         formatOptions: ts.FormatCodeSettings
-    ): ts.CodeAction[];
+    ): Array<ts.CodeAction | ts.CodeFixAction>;
     
     getDefinitionAtPosition?(
         context: TemplateContext,
         position: ts.LineAndCharacter
     ): ts.DefinitionInfo[];
+
+    getSignatureHelpItemsAtPosition?(
+        context: TemplateContext,
+        position: ts.LineAndCharacter
+    ): ts.SignatureHelpItems;
+
+    getOutliningSpans?(
+        context: TemplateContext
+    ): ts.OutliningSpan[];
 }